+++ lib/internal/main/eval_string.js
@@ -14,7 +14,7 @@ const {
   markBootstrapComplete
 } = require('internal/process/pre_execution');
 const { evalModule, evalScript } = require('internal/process/execution');
-const { addBuiltinLibsToObject } = require('internal/modules/cjs/helpers');
+const { addBuiltinLibsToObject } = require('internal/modules/helpers');
 
 const { getOptionValue } = require('internal/options');
