+++ lib/internal/modules/cjs/loader.js
@@ -103,7 +103,7 @@ const {
   makeRequireFunction,
   normalizeReferrerURL,
   stripBOM,
-} = require('internal/modules/cjs/helpers');
+} = require('internal/modules/helpers');
 const { getOptionValue } = require('internal/options');
 const preserveSymlinks = getOptionValue('--preserve-symlinks');
 const preserveSymlinksMain = getOptionValue('--preserve-symlinks-main');
