+++ lib/internal/modules/esm/translators.js
@@ -30,7 +30,7 @@ const {
   hasEsmSyntax,
   loadBuiltinModule,
   stripBOM,
-} = require('internal/modules/cjs/helpers');
+} = require('internal/modules/helpers');
 const {
   Module: CJSModule,
   cjsParseCache
