+++ lib/internal/source_map/source_map_cache.js
@@ -27,7 +27,7 @@ const { getOptionValue } = require('internal/options');
 const { IterableWeakMap } = require('internal/util/iterable_weak_map');
 const {
   normalizeReferrerURL,
-} = require('internal/modules/cjs/helpers');
+} = require('internal/modules/helpers');
 const { validateBoolean } = require('internal/validators');
 const { setMaybeCacheGeneratedSourceMap } = internalBinding('errors');
