+++ lib/internal/util/inspector.js
@@ -66,7 +66,7 @@ function installConsoleExtensions(commandLineApi) {
   if (commandLineApi.require) { return; }
   const { tryGetCwd } = require('internal/process/execution');
   const CJSModule = require('internal/modules/cjs/loader').Module;
-  const { makeRequireFunction } = require('internal/modules/cjs/helpers');
+  const { makeRequireFunction } = require('internal/modules/helpers');
   const consoleAPIModule = new CJSModule('<inspector console>');
   const cwd = tryGetCwd();
   consoleAPIModule.paths = ArrayPrototypeConcat(
