+++ lib/repl.js
@@ -101,7 +101,7 @@ const { BuiltinModule } = require('internal/bootstrap/loaders');
 const {
   makeRequireFunction,
   addBuiltinLibsToObject
-} = require('internal/modules/cjs/helpers');
+} = require('internal/modules/helpers');
 const {
   isIdentifierStart,
   isIdentifierChar
